Regenerated stats CSV for Experiment 2
Rewrote function to generate the CSV file for Experiment 2, which I must have removed at some point during tidying up. I then regenerated the CSV file, which now has greater decimal precision, resulting in a small changes in the decimals of the stats. So I've rerun the stats and these are now the canonical versions